+++ ninio-snmp/src/main/java/com/davfx/ninio/snmp/OidBerPacket.java
@@ -10,7 +10,7 @@ public final class OidBerPacket implements BerPacket {
 	private final ByteBuffer bb;
 
 	public OidBerPacket(Oid oid) {
-		int[] raw = oid.raw;
+		long[] raw = oid.raw;
 
 		if (raw.length < 2) {
 			throw new IllegalArgumentException();
@@ -20,7 +20,7 @@ public final class OidBerPacket implements BerPacket {
 		bb.put((byte) ((raw[1] + (raw[0] * 40)) & 0xFF));
 
 		for (int i = 2; i < raw.length; i++) {
-			int value = raw[i];
+			long value = raw[i];
 
 			int mask = ~0x80;
 			int bits = 0;
@@ -34,7 +34,7 @@ public final class OidBerPacket implements BerPacket {
 			}
 
 			while (bits >= 0) {
-				int b = (value >>> bits) & ~0x80;
+				long b = (value >>> bits) & ~0x80;
 				if (bits > 0) {
 					b |= 0x80; // Continuation bit
 				}
